fix parse_seq  when expecting ranges that are not required
There are certain cases when SP=0 and ISR=0, the parsing functions expect
a dash after the first number to get a range and throw an error condition if
they didn't.  Fixed the functions to handle this case, use the flat value
and continue to parse.

Also edited the sample config to use a similar fingerprint that actually
exists, the previous personality is no longer inside the nmap-os-db.