Fix for all lines except SEQ
This build currently fingerprints correctly on everyline except for the SEQ
line and occasionally the OPS line.  I believe they both have to do with the
personate of seq_tests[2] and seq_test[3], specifically that the mss value
we use is not differentiating between the two. 

The symptoms are: garbage SEQ lines except for TS= and when scanning
Windows Server 2003, O3= is the same as all other O#= when it should not.

Other than these two problems it appears that honeyd is responding to the
fingerprint tests correctly.