Fix timestamp field for sequence T ops and TS tests, bound range on ISN
TS tests are now returning correct values.
Bounding the range on the ISN using a max and min increment value is now
showing correct SP and ISR values regularly.  All fingerprints tested appear
to be being detected with 100% accuracy.  Although special cases are likely
to remain where this is not the case.