Fix for segfaults in the ARP splay tree
It turns out the broadcast MAC address was getting stuck into the MAC -> IP
splay tree. The MAC would get updated after an ARP request, but the tree
would still be using the old key. Once the ARP timeout hit, the IP address
would be deleted from the IP -> MAC tree, the MAC would be checked in the
MAC -> IP tree, but nothing would come up since it's key would originally
the broadcast and got changed. The req would be free'd, and next MAC -> IP
search would segfault because the entry was free'd in the expire.

Tl;dr summary: Fixes a periodic segfault