Make sure interface isn't null before trying to call honeyd_deliver_ethernet
This to keep honeyd from crashing if it's trying to send an ICMP error
to an IP that it doesn't know how to contact. Template will be default
with a NULL inter variable, and you can't start calling send_ethernet
functions in that case.

It's possible we should just not send anything at all if interface_fine_resp 
returns NULL?

Ticket #33
Ticket #35