SCRIPTS: Working on IPP UDP Response, only have one thing to fix
For some reason, there's one byte of extra, random data at the end of the
packet data being passed back to honyed by the script. It's not being 
appended from within the script, which is confusing, but wireshark sees it
and complains about it, despite the packet being an SNMP response that
it'll recognize. Work more on it tomorrow.

#50