SCRIPTS: More work on UDP IPP Responses
snmpscrape.py will snmpwalk an ip address, get as many of the MIB values
as it can, and store them in a file. The idea will be that, with this file of MIB
values, the ipp scripts will just have to parse and match a request OID
with one from the file, construct the correct value parameter 
(type byte, length byte, and value) and send it back, thus mimicing the 
response an actual printer will get. Idea's there, implementation isn't as of 
yet. 

#50