SCRIPTS: snmpscrape fully restructures snmpwalk results
Using the restructured file should be as easy as parsing for a response OID
matching the request id, appending the type value, calculating the length
of the response, and then hexlifying the response for the return packet.

To finish this script series, a couple things need to happen: 
I found that an SNMP agent responds to MIBs in a best-match manner; 
there are cases where the request will contain an OID that doesn't trace all
the way down to a leaf of the tree, and the expected behavior is that the
responding agent will return all values that can claim the given request as a
parent. This functionality is not within the script yet.

In addition, the ipp_udp_wrap.py needs to parse for multiple OIDs within 
the request segment, and respond to all in kind. I'm not sure how the 
responding agent is supposed to handle multiple-oid requests...Does he
respond to the OIDs all in one message, or does he respond to each request
OID in a separate packet? I'll have to research more. I hope that it's the 
former.

#50