SCRIPTS: Beginning work on SNMP request OID matching
Trying to find a way for the script to respond to a request with multiple 
packets per one request, if needed. creating an snmprequest.py script
to send a single snmp get-request packet to a specified ip address to test
whether multiple prints within a honeyd script will correlate to multiple
packets.

#50