SCRIPTS: Script responds properly to snmpwalk now
All of the values from the parsed file are parsed correctly to 
generate the desired output values, with the exception of some integer 
values that will be cleaned up in another commit. Going to start messing
around with 'find printer' and such to see if the honeypot will actually show
up, and then see if I can't make honeyd respond a little faster than it is
with the script data.

#50