SCRIPTS: Read filename for MIB file from generated configuration file
When one attaches the SNMP script to a node, a generated file is formed 
from the option subtree within the scripts.xml file. Note that this won't
happen if the snmpscrape script isn't run, as there's no mib file at that 
point; if one wishes to manually construct a file, it must be in the correct
syntax and it must have commensurate values within scripts.xml.

#50

Signed-off-by: Addison Waldow <addison.waldow@datasoft.com>